Added th:fragment for header Added collapse form to add routes Added very basic endpoint to receive new routes and save them

+++ src/main/kotlin/com/mwss/mytravelexpenses/model/Route.kt
@@ -17,6 +17,6 @@ open class Route {
     open var endLocation: String? = null
 
     @Column(name = "DISTANCE_IN_KM", nullable = false)
-    open var distanceInKm: Int = 0
+    open var distanceInKm: Double = 0.0
 
 }
